requirements.txt
flask
python-dotenv
flask-restful
flask-sqlalchemy
flask_marshmallow
sqlalchemy-filters
Faker
colorama